omni_core/requirements.txt
rich
psutil
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
